refactor: 커스텀 훅으로 기능 분리
- 메시지 데이터 관리하는 훅
- 메시지 스크롤 최하단으로 내려주는 훅
