+++ client/src/pages/MessagePage/MessageDetail/MessageDetailInner.tsx
@@ -1,14 +1,14 @@
 /** @jsxImportSource @emotion/react */
 
-import { useState, useEffect, useCallback, useRef, ChangeEvent, FormEvent } from 'react';
-import { useRecoilState, useRecoilValue } from 'recoil';
+import { useCallback } from 'react';
+import { useRecoilValue } from 'recoil';
 import { useNavigate } from 'react-router-dom';
 
 import { Button } from 'common';
-import { currentUserState, newMessageState } from 'store';
+import { currentUserState } from 'store';
 import { MessageDetailType, SingleMessageType } from 'types/message';
+import { useManageMessageData, useScrollBottom } from '../hooks';
 import { MessageElement } from './MessageElement';
-import { fetchSendMessage } from '../services';
 import { MessageTopBar } from '../MessageTopBar';
 
 import { messageDetailInputWrapperStyle, messageDetailListStyle, messageInputStyle, goBackButtonStyle } from './styles';
@@ -21,57 +21,19 @@ interface Props {
 }
 
 export const MessageDetailInner = ({ promise, userId }: Props) => {
-  const [currentMessageData, setCurrentMessageData] = useState<SingleMessageType[]>([]);
-  const [inputValue, setInputValue] = useState<string>('');
   const { id: currentUserID } = useRecoilValue(currentUserState);
-  const [newMessage, setNewMessage] = useRecoilState(newMessageState);
   const { toUsername, contents } = promise.read();
+  const { handleChangeInputValue, handleSubmitMessage, inputValue, currentMessageData } = useManageMessageData(
+    currentUserID ?? '',
+    contents,
+    userId
+  );
+  const { messageDetailListRef } = useScrollBottom(currentMessageData);
   const nav = useNavigate();
-
-  async function handleSubmitMessage(e: FormEvent<HTMLFormElement>) {
-    e.preventDefault();
-    if (inputValue.length < 1) return;
-    await fetchSendMessage(userId, inputValue)
-      .then(() => {
-        setCurrentMessageData((prevState) => [
-          ...prevState,
-          { from: currentUserID ?? '', content: inputValue, createdAt: new Date().toString() },
-        ]);
-      })
-      .then(() => {
-        setInputValue('');
-      });
-  }
-
-  useEffect(() => {
-    setCurrentMessageData(contents);
-  }, [contents]);
-
-  useEffect(() => {
-    if (!newMessage) return;
-    setCurrentMessageData((prevState) => [...prevState, newMessage]);
-    setNewMessage(null);
-  }, [newMessage]);
 
   const handleClickBackButton = useCallback(() => {
     nav('/message');
   }, []);
-
-  const messageDetailListRef = useRef<HTMLUListElement>(null);
-
-  function scrollToEnd() {
-    if (messageDetailListRef.current) {
-      messageDetailListRef.current.scrollTop = messageDetailListRef.current.scrollHeight;
-    }
-  }
-
-  function handleChangeInputValue(e: ChangeEvent<HTMLInputElement>) {
-    setInputValue(e.currentTarget.value);
-  }
-
-  useEffect(() => {
-    scrollToEnd();
-  }, [currentMessageData]);
 
   return (
     <>
