+++ client/src/pages/MessagePage/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useManageMessageData } from './useManageMessageData';
+export { useScrollBottom } from './useScrollBottom';
